fix(duration): accept media shorter than one second
The guard that turns unreadable metadata into media_metadata_missing
called to_i on a duration the analyzer reports as a rounded float, so
every value below 1.0 truncated to 0 and was treated as if ffprobe had
failed. A 0.5s clip validated with less_than: 2.seconds was rejected, and
with a misleading error at that.

to_f keeps the guard working for the cases it exists for: both nil and
the empty string a memoized-unavailable key round-trips to still give 0.0.

+++ lib/active_storage_validations/duration_validator.rb
@@ -30,7 +30,9 @@ module ActiveStorageValidations
           next
         end
 
-        if duration.to_i <= 0
+        # to_f, not to_i: durations are floats, and truncating would treat any
+        # file shorter than a second as unanalyzable.
+        if duration.to_f <= 0
           add_media_metadata_missing_error(record, attribute, attachable)
           next
         end
